Refactor zvol in to platform and platform independent bits
