+++ include/sys/Makefile.am
@@ -125,7 +125,8 @@ COMMON_H = \
 KERNEL_H = \
 	$(top_srcdir)/include/sys/zfs_ioctl.h \
 	$(top_srcdir)/include/sys/zfs_onexit.h \
-	$(top_srcdir)/include/sys/zvol.h
+	$(top_srcdir)/include/sys/zvol.h \
+	$(top_srcdir)/include/sys/zvol_impl.h
 
 USER_H =
