+++ module/os/linux/zfs/Makefile.in
@@ -33,3 +33,4 @@ $(MODULE)-objs += ../os/linux/zfs/zpl_file.o
 $(MODULE)-objs += ../os/linux/zfs/zpl_inode.o
 $(MODULE)-objs += ../os/linux/zfs/zpl_super.o
 $(MODULE)-objs += ../os/linux/zfs/zpl_xattr.o
+$(MODULE)-objs += ../os/linux/zfs/zvol_os.o
