+++ module/zfs/zfs_ioctl.c
@@ -4136,7 +4136,7 @@ static int
 zfs_ioc_rollback(const char *fsname, nvlist_t *innvl, nvlist_t *outnvl)
 {
 	zfsvfs_t *zfsvfs;
-	zvol_state_t *zv;
+	zvol_state_handle_t *zv;
 	char *target = NULL;
 	int error;
 
@@ -4861,7 +4861,7 @@ zfs_ioc_recv_impl(char *tofs, char *tosnap, char *origin, nvlist_t *recvprops,
 
 	if (error == 0) {
 		zfsvfs_t *zfsvfs = NULL;
-		zvol_state_t *zv = NULL;
+		zvol_state_handle_t *zv = NULL;
 
 		if (getzfsvfs(tofs, &zfsvfs) == 0) {
 			/* online recv */
@@ -7671,7 +7671,7 @@ _init(void)
 {
 	int error;
 
-	if ((error = -zvol_init()) != 0)
+	if ((error = zvol_init()) != 0)
 		return (error);
 
 	spa_init(FREAD | FWRITE);
